Token 만료 시 자동 로그아웃 구현 & 리뷰 수정 시 기존 내용 추가

+++ src/Pages/CrewPage/components/PhotoDetailModal.jsx
@@ -54,9 +54,11 @@ function PhotoDetailModal({ onClose, photoId, postId }) {
         <DeleteBtn
           type="button"
           onClick={() => {
-            dispatch(deleteCrewPhotos(id));
-            dispatch(deleteCrewPhoto(id));
-            onClose(modalRef);
+            if(window.confirm("정말 삭제하시겠어요?")){
+              dispatch(deleteCrewPhotos(id));
+              dispatch(deleteCrewPhoto(id));
+              onClose(modalRef);
+            }
           }}
         >
           삭제
